+++ mujoco_playground/_src/wrapper.py
@@ -93,18 +93,22 @@ def wrap_for_brax_training(
     randomization_fn: Optional[
         Callable[[mjx.Model], Tuple[mjx.Model, mjx.Model]]
     ] = None,
+    full_reset: bool = False,
 ) -> Wrapper:
   """Common wrapper pattern for all brax training agents.
 
   Args:
     env: environment to be wrapped
     vision: whether the environment will be vision based
-    num_vision_envs: number of environments the renderer should generate,
-      should equal the number of batched envs
+    num_vision_envs: number of environments the renderer should generate, should
+      equal the number of batched envs
     episode_length: length of episode
     action_repeat: how many repeated actions to take per step
     randomization_fn: randomization function that produces a vectorized model
       and in_axes to vmap over
+    full_reset: whether to call `env.reset` during `env.step` on done rather
+      than resetting to a cached first state. Setting full_reset=True may
+      increase wallclock time because it forces full resets to random states.
 
   Returns:
     An environment that is wrapped with Episode and AutoReset wrappers.  If the
@@ -118,24 +122,66 @@ def wrap_for_brax_training(
   else:
     env = BraxDomainRandomizationVmapWrapper(env, randomization_fn)
   env = brax_training.EpisodeWrapper(env, episode_length, action_repeat)
-  env = BraxAutoResetWrapper(env)
+  env = BraxAutoResetWrapper(env, full_reset=full_reset)
   return env
 
 
 class BraxAutoResetWrapper(Wrapper):
-  """Automatically resets Brax envs that are done."""
+  """Automatically resets Brax envs that are done.
+
+  If `full_reset` is disabled (default):
+    * the environment will reset to a cached first state.
+    * only data and obs are reset, not the environment info.
+
+  If `full_reset` is enabled:
+    * the environment will call env.reset during env.step on done.
+    * `full_reset` will thus incur a penalty in wallclock time depending on the
+      complexity of the reset function.
+    * info is fully reset, except for info under the key
+      `AutoResetWrapper_preserve_info`, which is passed through from the prior
+      step. This can be used for curriculum learning.
+
+  Attributes:
+    env: The wrapped environment.
+    full_reset: Whether to call `env.reset` during `env.step` on done.
+  """
+
+  def __init__(self, env: Any, full_reset: bool = False):
+    super().__init__(env)
+    self._full_reset = full_reset
+    self._info_key = 'AutoResetWrapper'
 
   def reset(self, rng: jax.Array) -> mjx_env.State:
-    state = self.env.reset(rng)
-    state.info['first_state'] = state.data
-    state.info['first_obs'] = state.obs
+    rng_key = jax.vmap(jax.random.split)(rng)
+    rng, key = rng_key[..., 0], rng_key[..., 1]
+    state = self.env.reset(key)
+    state.info[f'{self._info_key}_first_data'] = state.data
+    state.info[f'{self._info_key}_first_obs'] = state.obs
+    state.info[f'{self._info_key}_rng'] = rng
+    state.info[f'{self._info_key}_done_count'] = jp.zeros(
+        key.shape[:-1], dtype=int
+    )
     return state
 
   def step(self, state: mjx_env.State, action: jax.Array) -> mjx_env.State:
+    # grab the reset state.
+    reset_state = None
+    rng_key = jax.vmap(jax.random.split)(state.info[f'{self._info_key}_rng'])
+    reset_rng, reset_key = rng_key[..., 0], rng_key[..., 1]
+    if self._full_reset:
+      reset_state = self.reset(reset_key)
+      reset_data = reset_state.data
+      reset_obs = reset_state.obs
+    else:
+      reset_data = state.info[f'{self._info_key}_first_data']
+      reset_obs = state.info[f'{self._info_key}_first_obs']
+
     if 'steps' in state.info:
+      # reset steps to 0 if done.
       steps = state.info['steps']
       steps = jp.where(state.done, jp.zeros_like(steps), steps)
       state.info.update(steps=steps)
+
     state = state.replace(done=jp.zeros_like(state.done))
     state = self.env.step(state, action)
 
@@ -147,11 +193,25 @@ class BraxAutoResetWrapper(Wrapper):
         done = jp.reshape(done, [x.shape[0]] + [1] * (len(x.shape) - 1))
       return jp.where(done, x, y)
 
-    data = jax.tree.map(
-        where_done, state.info['first_state'], state.data
-    )
-    obs = jax.tree.map(where_done, state.info['first_obs'], state.obs)
-    return state.replace(data=data, obs=obs)
+    data = jax.tree.map(where_done, reset_data, state.data)
+    obs = jax.tree.map(where_done, reset_obs, state.obs)
+
+    next_info = state.info
+    done_count_key = f'{self._info_key}_done_count'
+    if self._full_reset and reset_state:
+      next_info = jax.tree.map(where_done, reset_state.info, state.info)
+      next_info[done_count_key] = state.info[done_count_key]
+
+      if 'steps' in next_info:
+        next_info['steps'] = state.info['steps']
+      preserve_info_key = f'{self._info_key}_preserve_info'
+      if preserve_info_key in next_info:
+        next_info[preserve_info_key] = state.info[preserve_info_key]
+
+    next_info[done_count_key] += state.done.astype(int)
+    next_info[f'{self._info_key}_rng'] = reset_rng
+
+    return state.replace(data=data, obs=obs, info=next_info)
 
 
 class BraxDomainRandomizationVmapWrapper(Wrapper):